refactor(dashboard): extract keyspace list card into shared StatsListCard
The card shell, sparkline, tooltip, and empty/loading/error chart states move
from apis/_components/api-list-card.tsx to components/stats-list-card with a
neutral two-series API (success/error buckets plus display labels); data
fetching stays with the caller. ApiListCard becomes a thin wrapper and the
keyspace list skeleton is replaced by the shared StatsListCardSkeleton.
