+++ web/apps/dashboard/components/stats-list-card/skeleton.tsx
@@ -1,14 +1,14 @@
 import { Skeleton } from "@unkey/ui";
 
-// Mirrors ApiListCard's layout exactly (same paddings, header line heights, and the
-// h-12 chart well) so swapping the list-loading skeleton for real cards causes no
+// Mirrors StatsListCard's layout exactly (same paddings, header line heights, and the
+// h-12 chart well) so swapping a list-loading skeleton for real cards causes no
 // layout shift.
-export function ApiCardSkeleton() {
+export function StatsListCardSkeleton({ subtitle = false }: { subtitle?: boolean }) {
   return (
     <div className="relative h-full p-5 flex flex-col border border-grayA-4 rounded-lg w-full gap-5">
       <div className="flex flex-col w-full gap-2 min-w-0">
         <Skeleton className="h-[14px] w-32" />
-        <Skeleton className="h-[12px] w-40" />
+        {subtitle ? <Skeleton className="h-[12px] w-40" /> : null}
       </div>
       <div className="mt-auto flex flex-col gap-3">
         <Skeleton className="h-12 w-full" />
